requirements.txt
Django==1.6.2
dj-database-url==0.3.0
dj-static==0.0.5
django-admin-sortable2==0.2.4
django-database-storage==0.1.2
django-summernote==0.5.5
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
